# internal/rangekey: sort keys by suffix once during iteration