… table_stats.go