During iteration over range keys, a stack of keyspan.FragmentIterators handle
various parts of range-key iteration: fragmenting, shadowing and defragmenting.
Previously, the keyspan.Span and keyspan.FragmentIterator strictly specified
the order of keys within a Span: always by Trailer, descending.

This created some unfortunate gymnastics during range-key iteration:
- To resolve shadowing (or "coalesce" range keys), the keyspan.Transform
  invoked by the keyspan.MergingIter) would sort the keys by suffix, and
  re-sort by Trailer.
- To determine whether or not two range key spans should be defragmented into a
  single Span, the range-key iterator would copy the keys into a buffer to sort
  by suffix.
- Lastly, when the range keys were surfaced to the *pebble.Iterator, they'd
  once again be sorted by suffix to present to the client.

This commit relaxes the keyspan.Span contract, allowing the Transform that
resolves shadowing to leave range keys in sorted order. This allows the
defragmenting iterator and the pebble.Iterator to avoid any additional sorting.

Benchmark results when rebased over #1824:
```
name                                                           old time/op    new time/op    delta
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward-10      2.31ms ± 6%    2.20ms ± 6%   -4.89%  (p=0.000 n=20+18)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-10     3.45ms ±14%    3.17ms ± 3%   -8.22%  (p=0.000 n=20+19)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward-10      2.20ms ±14%    1.97ms ± 3%  -10.48%  (p=0.000 n=19+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-10     2.83ms ± 4%    2.64ms ± 1%   -6.79%  (p=0.000 n=19+17)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward-10       978µs ± 1%     957µs ± 2%   -2.10%  (p=0.000 n=19+18)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-10     1.37ms ± 3%    1.38ms ± 5%     ~     (p=0.134 n=18+19)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward-10      151µs ± 1%     150µs ± 1%   -0.78%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-10     304µs ± 1%     311µs ± 1%   +2.31%  (p=0.000 n=20+19)

name                                                           old alloc/op   new alloc/op   delta
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward-10      2.88kB ± 1%    2.62kB ± 1%   -9.03%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-10     4.50kB ± 0%    4.22kB ± 1%   -6.38%  (p=0.000 n=16+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward-10      2.89kB ± 0%    2.62kB ± 1%   -9.42%  (p=0.000 n=14+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-10     4.47kB ± 1%    4.22kB ± 1%   -5.74%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward-10      2.86kB ± 0%    2.61kB ± 0%   -8.94%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-10     4.45kB ± 0%    4.19kB ± 0%   -5.72%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward-10     2.85kB ± 0%    2.59kB ± 0%   -8.98%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-10    4.45kB ± 0%    4.20kB ± 0%   -5.74%  (p=0.000 n=20+20)

name                                                           old allocs/op  new allocs/op  delta
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward-10        41.0 ± 0%      33.0 ± 0%  -19.51%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward-10       46.0 ± 0%      38.0 ± 0%  -17.39%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward-10        41.0 ± 0%      33.0 ± 0%  -19.51%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward-10       46.0 ± 0%      38.0 ± 0%  -17.39%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward-10        41.0 ± 0%      33.0 ± 0%  -19.51%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward-10       46.0 ± 0%      38.0 ± 0%  -17.39%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward-10       41.0 ± 0%      33.0 ± 0%  -19.51%  (p=0.000 n=20+20)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward-10      47.0 ± 0%      39.0 ± 0%  -17.02%  (p=0.000 n=20+20)
```